+++ README.md
@@ -2,5 +2,3 @@ sqflodo
 =======
 
 Todo list with [Flask](http://flask.pocoo.org/) and [SQLite](https://www.sqlite.org/)
-
-
