+++ sqflodo.py
@@ -2,45 +2,23 @@ import os.path, datetime
 from flask import Flask, render_template, request, url_for
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import Column, Integer, String, Date, Boolean, create_engine
-from flask_wtf import FlaskForm
-from wtforms import StringField, BooleanField, SubmitField
-from wtforms.fields.html5 import DateField
-from wtforms.validators import DataRequired
+from forms import *
 
 app = Flask(__name__)
-db_path = os.path.join(os.path.dirname(__file__), "tasks.db")
-db_uri = "sqlite:///{}".format(db_path)
-app.config.update(dict(
-        SQLALCHEMY_DATABASE_URI = db_uri,
-        SQLALCHEMY_TRACK_MODIFICATIONS = False,
-        SECRET_KEY = "secretkey",
-        WTF_CSRF_SECRET_KEY= "supersecretkey"
-        ))
+app.config.from_object('config')
 db = SQLAlchemy(app)
-
 
 # Models
 
 def _get_date():
     return datetime.datetime.now()
 
-class Tasks(db.Model):
-    key = db.Column(db.Integer, primary_key=True)
-    tdtask = db.Column(db.String(160), nullable=False)
-    createdday = db.Column(db.Date, default=datetime.datetime.utcnow)
-    lastday = db.Column(db.Date, default=datetime.datetime.utcnow)
-    complete = db.Column(db.Boolean, default=False)
-
-class AddTaskForm(FlaskForm):
-    newtask = StringField("newtask", validators=[DataRequired()])
-    duedate = DateField("duedate", format="%Y-%m-%d", validators=[DataRequired()])
-
-
 # Views
 @app.route("/", methods=["GET", "POST"])
 def index():
     alltasks = db.session.query(Tasks).all()
     atform = AddTaskForm(request.form)
+
     if request.method == "POST":
         newtask = request.form.get("newtask")
 
@@ -61,9 +39,8 @@ def index():
         db.session.add(row)
             
         db.session.commit()
-            
-    return render_template("index.html", mytasks = alltasks, atform = atform)
 
+    return render_template("index.html", mytasks = alltasks, atform = atform)
 
 if __name__ == "__main__":
     app.run(debug=True)
