+++ templates/base.html
@@ -18,7 +18,6 @@
 
 <main>
   {% block content %}{% endblock %}
-
 </main>
 
 </body>
